+++ Assets/Scripts/Installers/MenuInstaller.cs
@@ -1,0 +1,17 @@
+using UnityEngine;
+using Zenject;
+
+namespace Installers
+{
+    public class MenuInstaller: MonoInstaller
+    {
+        //[SerializeField] private PlayerInventory playerInventory;
+        [SerializeField] private WindowsManager windowsManager;
+        
+        public override void InstallBindings()
+        {
+            //Container.Bind<PlayerInventory>().FromInstance(playerInventory).AsSingle();
+            Container.Bind<WindowsManager>().FromInstance(windowsManager).AsSingle();
+        }
+    }
+}
