+++ Assets/Scripts/Installers/MenuInstaller.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 09c41a16dd394615a578b5f500503ea0
+timeCreated: 1679171792
