--- Assets/Scripts/Menu.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d1b35f5e02ba19f489482da7de3f2db3
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
