+++ Assets/Scripts/UI/Menu/ExitButton.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1a62ba3a4ec6af845aeadd5e63dfc965
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
