+++ Assets/Scripts/UI/Menu/OnClickButton.cs
@@ -1,0 +1,19 @@
+using UnityEngine;
+using UnityEngine.UI;
+
+namespace UI.Menu
+{
+    [RequireComponent(typeof(Button))]
+    public abstract class OnClickButton: MonoBehaviour
+    {
+        private Button _button;
+
+        private void Start()
+        {
+            _button = GetComponent<Button>();
+            _button.onClick.AddListener(OnClick);
+        }
+
+        protected abstract void OnClick();
+    }
+}
