+++ Assets/Scripts/UI/Windows Manager/WindowType.cs
@@ -2,6 +2,7 @@
 {
     public enum WindowType
     {
-        
+        Main,
+        Options
     }
 }
